Update README and create dataset.py
